refactored controllers using params_helper and edgecases_helper
