+++ app/controllers/api/v1/items_controller.rb
@@ -1,4 +1,5 @@
 class Api::V1::ItemsController < ApplicationController
+  include ParamsHelper
   def index
     items = Item.all.paginate(page: page_num, per_page: num_per_page)
     render json: Api::V1::ItemSerializer.new(items)
@@ -13,11 +14,8 @@ class Api::V1::ItemsController < ApplicationController
   end
 
   def update 
-    if valid_merchant_id
-      render json: Api::V1::ItemSerializer.new(Item.update(params[:id], item_params))
-    else 
-      render status: 404
-    end
+    render json: Api::V1::ItemSerializer.new(Item.update(params[:id], item_params)) if valid_merchant_id
+    render status: 404 if !valid_merchant_id
   end
 
   def destroy 
@@ -27,17 +25,6 @@ class Api::V1::ItemsController < ApplicationController
   end
   
 private
-  def item_params
-    params.require(:item).permit(:name, :description, :unit_price, :merchant_id)
-  end
-
-  def page_num
-    params[:page]? params[:page] : 1
-  end
-
-  def num_per_page
-    params[:per_page]? params[:per_page] : 20
-  end
 
   def valid_merchant_id
     Merchant.id_exist?(params[:merchant_id]) || !item_params[:merchant_id]
