+++ app/controllers/api/v1/merchant_items_controller.rb
@@ -1,4 +1,5 @@
 class Api::V1::MerchantItemsController < ApplicationController
+
 before_action :set_merchant
   def index
   	render json: Api::V1::ItemSerializer.new(@merchant.items)
