+++ app/controllers/api/v1/merchants_controller.rb
@@ -1,4 +1,5 @@
 class Api::V1::MerchantsController < ApplicationController
+  include ParamsHelper
   def index
     merchants = Merchant.all.paginate(page: page_num, per_page: num_per_page)
   	render json: Api::V1::MerchantSerializer.new(merchants)
@@ -8,12 +9,4 @@ class Api::V1::MerchantsController < ApplicationController
   	render json: Api::V1::MerchantSerializer.new(Merchant.find(params[:id]))
   end
 
-  private
-  def page_num
-    params[:page]? params[:page] : 1
-  end
-
-  def num_per_page
-    params[:per_page]? params[:per_page] : 20
-  end
 end
