+++ app/controllers/api/v1/search_merchants_controller.rb
@@ -1,48 +1,19 @@
 class Api::V1::SearchMerchantsController < ApplicationController
+  include ParamsHelper, EdgeCasesHelper
   def index
-    if edge_cases_conditions
-      edge_case_resposne
-    elsif name_search
-  	  render json: Api::V1::MerchantSerializer.new(Merchant.search_all_by_name(params[:name])) 
-    end
+    edge_case_resposne if edge_cases_conditions
+  	render json: Api::V1::MerchantSerializer.new(Merchant.search_all_by_name(params[:name])) if !edge_cases_conditions
+    
   end
 
   def show
     if edge_cases_conditions
       edge_case_resposne
-    elsif name_search
+    else
       merchant = Merchant.search_one_by_name(params[:name])
       render json: Api::V1::MerchantSerializer.new(merchant), status: :ok if merchant
       render json: {data: {error: 'Merchant not found'}} if !merchant
     end 
   end
 
-  private 
-    def name_search
-      params[:name] && params[:name] != ''
-    end
-
-    def no_params
-      !params[:name]
-    end
-
-    def empty_params
-       params[:name] ==''
-    end
-
-    def edge_cases_conditions
-      no_params || empty_params
-    end
-
-    def edge_case_resposne
-      return render json: Api::V1::ErrorSerializer.format_error(error_messages[:missing_param]), status: 400 if no_params 
-      return render json: Api::V1::ErrorSerializer.format_error(error_messages[:empty_param]), status: 400 if empty_params 
-    end
-
-    def error_messages
-      messages = Hash.new 
-      messages[:missing_param]='Parameter cannot be missing'
-      messages[:empty_param] = 'Parameter cannot be empty'
-      messages
-    end
 end
